Continue adding room functionality
Manage CPUs menu now contains a helpful message when there are no CPUs.
Only the host of a room can now modify settings and start the room.
Spectating half works. They remain in the room, but are unable to watch the game once it starts for unknown reasons.

Simplified some code structures.