Start work on guide
Set up Vue component for guide pages. Should be pretty smooth to add new ones.
Basic framework for simulator. Still needs a lot of work.
Reworked the routing, both on server-side and client-side. Organized some of the file structure to reflect this.
Removed the 'limited' javascript file. It might make loading times slightly worse on the static pages, but it's way simpler to deal with.
Fixed case-sensitive issue with image filenames.